consolidates patches
