+++ SolastaCommunityExpansion/Patches/GameUiLocation/UnityModManagerUIPatcher.cs
@@ -4,7 +4,7 @@ using System.Threading.Tasks;
 using HarmonyLib;
 using UnityModManagerNet;
 
-namespace SolastaCommunityExpansion.Patches.ModUi
+namespace SolastaCommunityExpansion.Patches.GameUiLocation
 {
     [HarmonyPatch(typeof(UnityModManager.UI), "ToggleWindow", new Type[] { typeof(bool) })]
     [SuppressMessage("Minor Code Smell", "S101:Types should be named in PascalCase", Justification = "Patch")]
